backend/requirements.txt
flask==1.1.2
flask-cors==3.0.9
paho-mqtt==1.5.1
gunicorn==20.0.4
